+++ projects/frontend/es-core/src/RecalboxConf.cpp
@@ -88,6 +88,7 @@ DefineSystemGetterSetterImplementation(Rewind, bool, Bool, sSystemRewind, GetGlo
 DefineSystemGetterSetterImplementation(AutoSave, bool, Bool, sSystemAutoSave, GetGlobalAutoSave())
 DefineSystemGetterSetterImplementation(Shaders, std::string, String, sSystemShaders, GetGlobalShaders())
 DefineSystemGetterSetterImplementation(ShaderSet, std::string, String, sSystemShaderSet, GetGlobalShaderSet())
+DefineSystemGetterSetterImplementation(VideoMode, std::string, String, sSystemVideoMode, "")
 
 DefineEmulationStationSystemGetterSetterImplementation(FilterAdult, bool, Bool, sSystemFilterAdult, false)
 DefineEmulationStationSystemGetterSetterImplementation(FlatFolders, bool, Bool, sSystemFlatFolders, false)